requirements.txt
# Python dependencies
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
# pySteamSignIn==1.1.1
duckduckgo-search==4.1.1
beautifulsoup4==4.12.2
lxml==5.1.0
openai==1.12.0  # For OpenRouter API (uses OpenAI SDK-compatible interface)
